Move FontInfo and GlyphMetrics structs to Font header
